requirements.txt
torch==2.5.1+cpu
torchvision==0.20.1
albumentations==1.3.0
opencv-python==4.8.0.76
numpy==1.24.0
Pillow==10.0.0
tqdm==4.64.0